Add files via upload
Followed this code from a youtube video: https://www.youtube.com/watch?v=TGSDz-_dNhI&t=914s
My idea is if we want to have difent pages we have to create different routes: e.g route("id") will display the login/register page
route("exercise") will display the exercises 
route("food") will display the food items 
and so on and so forth 

i used intelliJ IDEA to develop this, as so did the dude in the youtube video. 
still needs some work but will work on it in 2 hours or so, need to take a nap cuz i wouke uo at 3am to start this.

def think we need to create the routes so we can switch from page to page.